Expose core metadata validation functions in public headers
